requirements.txt
Django==4.1
djangorestframework==3.13.1
django-filter==22.1
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1